api: remove deprecated dashboard pairs past window
